+++ CHANGES.md
@@ -2,6 +2,10 @@
 Changes
 =======
 
+- 1.7.1
+    - IMPROVEMENT: improve performance of preview generation by switching from Jimp to Electron NativeImage
+    - IMPROVEMENT: fix NDI video frame generation on big endian CPU based computers
+
 - 1.7.0
     - IMPROVEMENT: support particular profile with the option "--profile=<id>|<directory>"
     - IMPROVEMENT: support particular auto-imported/exported external configuration with option "--config=<file>"